Improve testability
