+++ Cargo.toml
@@ -4,15 +4,16 @@ version = "0.1.1"
 edition = "2021"
 
 [dependencies]
-decon-spf = "0.3.3"
-trust-dns-resolver = { version = "0.23.2", features = ["tokio-runtime"] }
-futures = "0.3.31"
-tokio = { version = "1.44.2", features = ["full"] }
-axum = "0.8.3"
-serde = { version = "1.0.219", features = ["derive"] }
 anyhow = "1.0.98"
-http = "1.3.1"
+async-trait = "0.1"
+axum = "0.8.3"
 chrono = "0.4.40"
+decon-spf = "0.3.3"
+futures = "0.3.31"
+http = "1.3.1"
+serde = { version = "1.0.219", features = ["derive"] }
+tokio = { version = "1.44.2", features = ["full"] }
+trust-dns-resolver = { version = "0.23.2", features = ["tokio-runtime"] }
 
 [profile.release]
 lto = true
